Learning day endpoints added

+++ src/main/java/mif/vu/ikeea/Entity/Repository/TopicRepository.java
@@ -3,5 +3,5 @@ package mif.vu.ikeea.Entity.Repository;
 import mif.vu.ikeea.Entity.Topic;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TopicRepository extends CrudRepository<Topic, Integer> {
+public interface TopicRepository extends CrudRepository<Topic, Long> {
 }
